campusvirtual/requirements/base.txt
asgiref==3.3.4
astroid==2.11.7
Babel==2.9.1
Django==3.2
django-crum==0.7.9
django-phonenumber-field==5.1.0
django-widget-tweaks==1.4.8
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
phonenumbers==8.12.23
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
python-dateutil==2.8.1
django-filter==2.4.0
django-bootstrap-form==3.4
numpy==1.21.2
pandas==1.3.3
openpyxl==3.0.9
django-sequences==2.7
qrcode==7.3.1
pillow==9.2.0
djangorestframework==3.14.0
django-soft-delete==0.9.21